requirements.txt
numpy==1.21.3
opencv_python_headless==4.5.4.60
pyzbar==0.1.8
streamlit==1.1.0
Pillow==8.4.0
